Move AuthComponents to their own file
